consumer/protocol: add Checkpoint message; recovery log as optional
* Recovery logs are now optional on ShardSpecs, and are not needed if an
external store is used.

* Checkpoint bundles read-through offsets, message producer states, and
acknowledgement intents within a single message.

* TAILING redefined => STANDBY, which is applicable for a shard not having
a recovery log.

* StatRequest takes a new ReadThrough field, and StatResponse returns
ReadThrough and PublishAt. These fields allow authors of APIs built atop
consumers to offer "read your writes" consistency.

Issue #210
