+++ consumer/protocol/shard_spec_extensions_test.go
@@ -37,7 +37,7 @@ func (s *SpecSuite) TestShardSpecValidationCases(c *gc.C) {
 		Id:                "bad id",
 		Sources:           nil,
 		RecoveryLogPrefix: "bad prefix",
-		HintPrefix:        "",
+		HintPrefix:        "not empty",
 		HintBackups:       -1,
 		MaxTxnDuration:    +0,
 		MinTxnDuration:    -1,
@@ -51,10 +51,10 @@ func (s *SpecSuite) TestShardSpecValidationCases(c *gc.C) {
 		{Journal: "journal 2"},
 		{Journal: "journal/1", MinOffset: -1},
 	}
-	c.Check(spec.Validate(), gc.ErrorMatches, `RecoveryLog: not a valid token \(bad prefix/a-shard-id\)`)
+	c.Check(spec.Validate(), gc.ErrorMatches, `RecoveryLogPrefix: not a valid token \(bad prefix/a-shard-id\)`)
 	spec.RecoveryLogPrefix = ""
-	c.Check(spec.Validate(), gc.ErrorMatches, `expected RecoveryLogPrefix`)
-	spec.RecoveryLogPrefix = "recovery/logs"
+	c.Check(spec.Validate(), gc.ErrorMatches, `invalid non-empty HintPrefix with empty RecoveryLogPrefix \(not empty\)`)
+	spec.HintPrefix, spec.RecoveryLogPrefix = "", "recovery/logs"
 	c.Check(spec.Validate(), gc.ErrorMatches, `HintPrefix is not an absolute, clean, non-directory path \(\)`)
 	spec.HintPrefix = "relative/path"
 	c.Check(spec.Validate(), gc.ErrorMatches, `HintPrefix is not an absolute, clean, non-directory path \(relative/path\)`)
@@ -107,7 +107,7 @@ func (s *SpecSuite) TestShardSpecRoutines(c *gc.C) {
 	var all = keyspace.KeyValues{asn, {Decoded: allocator.Assignment{Slot: 0, AssignmentValue: primaryStatus}}}
 
 	c.Check(spec.IsConsistent(asn, all), gc.Equals, false)
-	status.Code = ReplicaStatus_TAILING
+	status.Code = ReplicaStatus_STANDBY
 	c.Check(spec.IsConsistent(asn, all), gc.Equals, true)
 	status.Code = ReplicaStatus_PRIMARY
 	c.Check(spec.IsConsistent(asn, all), gc.Equals, true)
@@ -210,7 +210,7 @@ func (s *SpecSuite) TestReplicaStatusValidationCases(c *gc.C) {
 	var status = ReplicaStatus{Code: -1}
 
 	c.Check(status.Validate(), gc.ErrorMatches, `Code: invalid code \(-1\)`)
-	status.Code = ReplicaStatus_TAILING
+	status.Code = ReplicaStatus_STANDBY
 	c.Check(status.Validate(), gc.IsNil)
 
 	status.Errors = []string{"error!"}
@@ -226,166 +226,15 @@ func (s *SpecSuite) TestReplicaStatusReduction(c *gc.C) {
 	var status = &ReplicaStatus{Code: ReplicaStatus_IDLE}
 
 	// Code is increased (only) by reducing with a ReplicaStatus of greater Code value.
-	status.Reduce(&ReplicaStatus{Code: ReplicaStatus_TAILING})
-	c.Check(status, gc.DeepEquals, &ReplicaStatus{Code: ReplicaStatus_TAILING})
+	status.Reduce(&ReplicaStatus{Code: ReplicaStatus_STANDBY})
+	c.Check(status, gc.DeepEquals, &ReplicaStatus{Code: ReplicaStatus_STANDBY})
 	status.Reduce(&ReplicaStatus{Code: ReplicaStatus_BACKFILL})
-	c.Check(status, gc.DeepEquals, &ReplicaStatus{Code: ReplicaStatus_TAILING})
+	c.Check(status, gc.DeepEquals, &ReplicaStatus{Code: ReplicaStatus_STANDBY})
 
 	// Multiple errors are accumulated.
 	status.Reduce(&ReplicaStatus{Code: ReplicaStatus_FAILED, Errors: []string{"err-1"}})
 	status.Reduce(&ReplicaStatus{Code: ReplicaStatus_FAILED, Errors: []string{"err-2"}})
 	c.Check(status, gc.DeepEquals, &ReplicaStatus{Code: ReplicaStatus_FAILED, Errors: []string{"err-1", "err-2"}})
-}
-
-func (s *SpecSuite) TestStatRequestValidationCases(c *gc.C) {
-	var req = StatRequest{
-		Header: badHeaderFixture(),
-		Shard:  "invalid shard",
-	}
-	c.Check(req.Validate(), gc.ErrorMatches, `Header.Etcd: invalid ClusterId .*`)
-	req.Header.Etcd.ClusterId = 1234
-	c.Check(req.Validate(), gc.ErrorMatches, `Shard: not a valid token \(invalid shard\)`)
-	req.Shard = "valid-shard"
-
-	c.Check(req.Validate(), gc.IsNil)
-}
-
-func (s *SpecSuite) TestStatResponseValidationCases(c *gc.C) {
-	var resp = StatResponse{
-		Status: 9101,
-		Header: *badHeaderFixture(),
-		Offsets: map[pb.Journal]int64{
-			"invalid journal": 123,
-		},
-	}
-	c.Check(resp.Validate(), gc.ErrorMatches, `Status: invalid status \(9101\)`)
-	resp.Status = Status_OK
-	c.Check(resp.Validate(), gc.ErrorMatches, `Header.Etcd: invalid ClusterId .*`)
-	resp.Header.Etcd.ClusterId = 1234
-	c.Check(resp.Validate(), gc.ErrorMatches, `Offsets\[invalid journal\]: not a valid token \(invalid journal\)`)
-	delete(resp.Offsets, "invalid journal")
-	resp.Offsets["a/journal"] = -456
-	c.Check(resp.Validate(), gc.ErrorMatches, `Offsets\[a/journal\]: invalid offset \(-456; expected >= 0\)`)
-	resp.Offsets["a/journal"] = 789
-
-	c.Check(resp.Validate(), gc.IsNil)
-}
-
-func (s *SpecSuite) TestListRequestValidationCases(c *gc.C) {
-	var req = ListRequest{
-		Selector: pb.LabelSelector{
-			Include: pb.LabelSet{Labels: []pb.Label{{Name: "a invalid name", Value: "foo"}}},
-			Exclude: pb.MustLabelSet("id", "bar"),
-		},
-	}
-	c.Check(req.Validate(), gc.ErrorMatches,
-		`Selector.Include.Labels\[0\].Name: not a valid token \(a invalid name\)`)
-	req.Selector.Include.Labels[0].Name = "a-valid-name"
-
-	c.Check(req.Validate(), gc.IsNil)
-}
-
-func (s *SpecSuite) TestListResponseValidationCases(c *gc.C) {
-	var resp = ListResponse{
-		Status: 9101,
-		Header: *badHeaderFixture(),
-		Shards: []ListResponse_Shard{
-			{
-				ModRevision: 0,
-				Spec: ShardSpec{
-					Id:                "a invalid id",
-					Sources:           []ShardSpec_Source{{Journal: "a/journal"}},
-					RecoveryLogPrefix: "a/log/prefix",
-					HintPrefix:        "/a/hint/prefix",
-					MaxTxnDuration:    1,
-				},
-				Route:  pb.Route{Primary: 0},
-				Status: nil,
-			},
-		},
-	}
-
-	c.Check(resp.Validate(), gc.ErrorMatches, `Status: invalid status \(9101\)`)
-	resp.Status = Status_OK
-	c.Check(resp.Validate(), gc.ErrorMatches, `Header.Etcd: invalid ClusterId .*`)
-	resp.Header.Etcd.ClusterId = 1234
-	c.Check(resp.Validate(), gc.ErrorMatches, `Shards\[0\].Spec.Id: not a valid token \(.*\)`)
-	resp.Shards[0].Spec.Id = "a-valid-id"
-	c.Check(resp.Validate(), gc.ErrorMatches, `Shards\[0\]: invalid ModRevision \(0; expected > 0\)`)
-	resp.Shards[0].ModRevision = 1
-	c.Check(resp.Validate(), gc.ErrorMatches, `Shards\[0\].Route: invalid Primary .*`)
-	resp.Shards[0].Route.Members = []pb.ProcessSpec_ID{{Zone: "zone", Suffix: "suffix"}}
-	c.Check(resp.Validate(), gc.ErrorMatches, `Shards\[0\]: length of Route.Members and Status are not equal \(1 vs 0\)`)
-	resp.Shards[0].Status = []ReplicaStatus{{Code: ReplicaStatus_TAILING}}
-
-	c.Check(resp.Validate(), gc.IsNil)
-}
-
-func (s *SpecSuite) TestApplyRequestValidationCases(c *gc.C) {
-	var req = ApplyRequest{
-		Changes: []ApplyRequest_Change{
-			{
-				ExpectModRevision: -1,
-				Upsert: &ShardSpec{
-					Id:                "invalid id",
-					Sources:           []ShardSpec_Source{{Journal: "a/journal"}},
-					RecoveryLogPrefix: "a/log/prefix",
-					HintPrefix:        "/a/hint/prefix",
-					MaxTxnDuration:    1,
-				},
-				Delete: "another-id",
-			},
-			{
-				ExpectModRevision: 0,
-				Delete:            "another invalid id",
-			},
-			{
-				ExpectModRevision: 1,
-			},
-		},
-	}
-
-	c.Check(req.Validate(), gc.ErrorMatches, `Changes\[0\]: both Upsert and Delete are set \(expected exactly one\)`)
-	req.Changes[0].Delete = ""
-	c.Check(req.Validate(), gc.ErrorMatches, `Changes\[0\].Upsert.Id: not a valid token \(invalid id\)`)
-	req.Changes[0].Upsert.Id = "a-valid-id"
-	c.Check(req.Validate(), gc.ErrorMatches, `Changes\[0\]: invalid ExpectModRevision \(-1; expected >= 0\)`)
-	req.Changes[0].ExpectModRevision = 0
-	c.Check(req.Validate(), gc.ErrorMatches, `Changes\[1\].Delete: not a valid token \(another invalid id\)`)
-	req.Changes[1].Delete = "other-valid-id"
-	c.Check(req.Validate(), gc.ErrorMatches, `Changes\[1\]: invalid ExpectModRevision \(0; expected > 0\)`)
-	req.Changes[1].ExpectModRevision = 1
-	c.Check(req.Validate(), gc.ErrorMatches, `Changes\[2\]: neither Upsert nor Delete are set \(expected exactly one\)`)
-	req.Changes[2].Delete = "yet-another-valid-id"
-
-	c.Check(req.Validate(), gc.IsNil)
-}
-
-func (s *SpecSuite) TestApplyResponseValidationCases(c *gc.C) {
-	var resp = ApplyResponse{
-		Status: 9101,
-		Header: *badHeaderFixture(),
-	}
-
-	c.Check(resp.Validate(), gc.ErrorMatches, `Status: invalid status \(9101\)`)
-	resp.Status = Status_OK
-	c.Check(resp.Validate(), gc.ErrorMatches, `Header.Etcd: invalid ClusterId .*`)
-	resp.Header.Etcd.ClusterId = 1234
-
-	c.Check(resp.Validate(), gc.IsNil)
-}
-
-func badHeaderFixture() *pb.Header {
-	return &pb.Header{
-		ProcessId: pb.ProcessSpec_ID{Zone: "zone", Suffix: "name"},
-		Route:     pb.Route{Primary: 0, Members: []pb.ProcessSpec_ID{{Zone: "zone", Suffix: "name"}}},
-		Etcd: pb.Header_Etcd{
-			ClusterId: 0, // ClusterId is invalid, but easily fixed up.
-			MemberId:  34,
-			Revision:  56,
-			RaftTerm:  78,
-		},
-	}
 }
 
 var _ = gc.Suite(&SpecSuite{})
